chore: rename component folders to camelCase
