+++ app/src/features/fullPicture/FullPicture.tsx
